Moved Parameters out of package.scala
You can’t reach it from within Java
